+++ src/test/java/chess/domain/piece/DirectionTest.java
@@ -1,8 +1,10 @@
-package chess.domain.board;
+package chess.domain.piece;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.List;
+import chess.domain.board.Coordinate;
+import chess.domain.piece.Direction;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
